requirements.txt
boto3
monai-weekly==1.2.dev2313
nibabel
numpy
pandas
Pillow
SimpleITK
six
torch
torchaudio
torchvision
s3fs
torch-lr-finder
scikit-learn
scikit-multilearn
pycox
lifelines
medcam